refactor(gir): Consolidate and update introspected class imports
- Merged multiple class-related functionalities into a new `introspected-classes.ts` file, enhancing organization and maintainability.
- Updated import statements across various files to reference the new consolidated class definitions, improving clarity.
- Removed deprecated `IntrospectedClass`, `IntrospectedClassFunction`, and related classes, streamlining the codebase.
- Enhanced type safety by ensuring all class-related types are consistently imported from the new module.
