+++ packages/cli/package.json
@@ -6,14 +6,14 @@
   "module": "src/index.ts",
   "type": "module",
   "bin": {
-    "ts-for-gir": "./lib/start.js"
+    "ts-for-gir": "./src/start.ts"
   },
   "engines": {
     "node": ">=22"
   },
   "scripts": {
-    "start": "yarn node --max_old_space_size=9216 lib/start.js",
-    "start:dev": "node --experimental-specifier-resolution=node --experimental-strip-types --experimental-transform-types --no-warnings src/start.ts",
+    "start:prod": "yarn node --max_old_space_size=9216 lib/start.js",
+    "start": "node --experimental-specifier-resolution=node --experimental-strip-types --experimental-transform-types --no-warnings src/start.ts",
     "build": "node --experimental-specifier-resolution=node --experimental-strip-types --experimental-transform-types --no-warnings esbuild.ts",
     "chmod": "chmod +x ./lib/start.js",
     "build:ts": "tsc",
