+++ packages/cli/src/start.ts
@@ -1,4 +1,4 @@
-#!/usr/bin/env node
+#!/usr/bin/env node --experimental-specifier-resolution=node --experimental-strip-types --experimental-transform-types --no-warnings
 import yargs from 'yargs'
 import { hideBin } from 'yargs/helpers'
