+++ packages/generator-typescript/src/package-data-parser.ts
@@ -1,8 +1,13 @@
 import { parseStringPromise } from 'xml2js'
 import { readFile } from 'fs/promises'
-import { resolve } from 'path'
+import { dirname, resolve } from 'path'
 import { Logger, splitModuleName, cleanString } from '@ts-for-gir/lib'
-import { __dirname } from './utils.ts'
+import { fileURLToPath } from 'url'
+
+// Get __filename on ESM
+const __filename = fileURLToPath(import.meta.url)
+// Get __dirname on ESM, resolve to the root directory of this package
+export const __dirname = resolve(dirname(__filename), '..')
 
 import type {
     OptionsGeneration,
