+++ packages/lib/src/constants.ts
@@ -1,5 +1,4 @@
 import { LibraryVersion } from './library-version.ts'
-import { __dirname } from './utils/path.ts'
 import { readJsonFile } from './utils/files.ts'
 import { join } from 'path'
 
@@ -15,7 +14,7 @@ export const PACKAGE = await readJsonFile<{
     license: string
     homepage: string
     author: string
-}>(join(__dirname, './package.json'))
+}>(join(process.cwd(), './package.json'))
 
 export const APP_NAME = 'ts-for-gir'
 export const APP_USAGE = 'TypeScript type definition generator for GObject introspection GIR files'
