+++ packages/lib/src/generators/dts/gobject.ts
@@ -1,6 +1,6 @@
 import { GirDirection } from "@gi.ts/parser";
 import { NativeType, NeverType } from "../../gir.ts";
-import { IntrospectedClassFunction } from "../../gir/class-function.ts";
+import { IntrospectedClassFunction } from "../../gir/introspected-classes.ts";
 import { IntrospectedFunctionParameter } from "../../gir/parameter.ts";
 import { IntrospectedNamespace } from "../../gir/namespace.ts";
