+++ packages/lib/src/generators/generator.ts
@@ -1,7 +1,7 @@
 import type { IntrospectedNamespace } from "../gir/namespace.ts";
-import type { IntrospectedClass } from "../gir/introspected-class.ts";
+import type { IntrospectedClass } from "../gir/introspected-classes.ts";
 import type { IntrospectedRecord } from "../gir/record.ts";
-import type { IntrospectedInterface } from "../gir/interface.ts";
+import type { IntrospectedInterface } from "../gir/introspected-classes.ts";
 import type { IntrospectedConstant } from "../gir/const.ts";
 import type { IntrospectedEnum } from "../gir/enum.ts";
 import type { IntrospectedError } from "../gir/error.ts";
@@ -13,10 +13,10 @@ import type { IntrospectedConstructor } from "../gir/constructor.ts";
 import type { IntrospectedDirectAllocationConstructor } from "../gir/direct-allocation-constructor.ts";
 import type { IntrospectedFunction } from "../gir/function.ts";
 import type { IntrospectedCallback } from "../gir/callback.ts";
-import type { IntrospectedClassCallback } from "../gir/class-callback.ts";
-import type { IntrospectedClassFunction } from "../gir/class-function.ts";
-import type { IntrospectedStaticClassFunction } from "../gir/static-function.ts";
-import type { IntrospectedVirtualClassFunction } from "../gir/virtual-function.ts";
+import type { IntrospectedClassCallback } from "../gir/introspected-classes.ts";
+import type { IntrospectedClassFunction } from "../gir/introspected-classes.ts";
+import type { IntrospectedStaticClassFunction } from "../gir/introspected-classes.ts";
+import type { IntrospectedVirtualClassFunction } from "../gir/introspected-classes.ts";
 import type { IntrospectedAlias } from "../gir/alias.ts";
 import type { TypeExpression } from "../gir.ts";
 import type { OptionsGeneration } from "../types/options-generation.ts";
