+++ packages/lib/src/gir-module.ts
@@ -25,18 +25,17 @@ import type { GirNSMember } from './gir/namespace.ts'
 import { IntrospectedAlias } from './gir/alias.ts'
 import { IntrospectedBase } from './gir/introspected-base.ts'
 import type { IntrospectedNamespaceMember } from './gir/introspected-namespace-member.ts'
-import { IntrospectedBaseClass } from './gir/introspected-base-class.ts'
-import { IntrospectedClass } from './gir/introspected-class.ts'
+import { IntrospectedBaseClass } from './gir/introspected-classes.ts'
+import { IntrospectedClass } from './gir/introspected-classes.ts'
 import { IntrospectedRecord } from './gir/record.ts'
-import { IntrospectedInterface } from './gir/interface.ts'
+import { IntrospectedInterface } from './gir/introspected-classes.ts'
 import { IntrospectedConstant } from './gir/const.ts'
 import { IntrospectedEnum } from './gir/enum.ts'
 import { IntrospectedError } from './gir/error.ts'
 import { IntrospectedFunction } from './gir/function.ts'
 import type { IntrospectedFunctionParameter } from './gir/parameter.ts'
-import type { IntrospectedClassFunction } from './gir/class-function.ts'
+import type { IntrospectedClassFunction, IntrospectedClassCallback } from './gir/introspected-classes.ts'
 import { IntrospectedCallback } from './gir/callback.ts'
-import { IntrospectedClassCallback } from './gir/class-callback.ts'
 import type { NSRegistry } from './gir/registry.ts'
 import { isPrimitiveType } from './utils/types.ts'
 import type { GirVisitor } from './visitor.ts'
