+++ packages/lib/src/gir/index.ts
@@ -1,9 +1,9 @@
 export * from "./alias.ts";
 export * from "./callback.ts";
-export * from "./class-callback.ts";
-export * from "./class-function.ts";
-export * from "./direct-allocation-constructor.ts";
+export * from "./introspected-base.ts";
 export * from "./function.ts";
+export * from "./introspected-classes.ts";
+export * from "./direct-allocation-constructor.ts";
 export * from "./constructor.ts";
 export * from "./const.ts";
 export * from "./data.ts";
@@ -12,11 +12,7 @@ export * from "./enum.ts";
 export * from "./error.ts";
 export * from "./generics.ts";
 export * from "./gir-complex-record.ts";
-export * from "./interface.ts";
-export * from "./introspected-base-class.ts";
-export * from "./introspected-base.ts";
 export * from "./introspected-class-member.ts";
-export * from "./introspected-class.ts";
 export * from "./introspected-namespace-member.ts";
 export * from "./namespace.ts";
 export * from "./parameter.ts";
@@ -25,5 +21,3 @@ export * from "./property.ts";
 export * from "./record.ts";
 export * from "./registry.ts";
 export * from "./signal.ts";
-export * from "./static-function.ts";
-export * from "./virtual-function.ts";
