+++ packages/lib/src/types/class.ts
@@ -3,13 +3,13 @@ import type {
 } from "../gir.ts";
 import type {
     IntrospectedClassFunction,
-} from "../gir/class-function.ts";
+} from "../gir/introspected-classes.ts";
 import type { IntrospectedConstructor } from "../gir/constructor.ts";
-import type { IntrospectedClassCallback } from "../gir/class-callback.ts";
+import type { IntrospectedClassCallback } from "../gir/introspected-classes.ts";
 import type { IntrospectedProperty, IntrospectedField } from "../gir/property.ts";
-import type { IntrospectedBaseClass } from "../gir/introspected-base-class.ts";
-import type { IntrospectedInterface } from "../gir/interface.ts";
-import type { IntrospectedClass } from "../gir/introspected-class.ts";
+import type { IntrospectedBaseClass } from "../gir/introspected-classes.ts";
+import type { IntrospectedInterface } from "../gir/introspected-classes.ts";
+import type { IntrospectedClass } from "../gir/introspected-classes.ts";
 import type { IntrospectedRecord } from "../gir/record.ts";
 
 export interface ClassDefinition {
