+++ packages/lib/src/utils/type-resolution.ts
@@ -1,5 +1,5 @@
 import type { IntrospectedNamespace } from "../gir/namespace.ts";
-import { IntrospectedBaseClass } from "../gir/introspected-base-class.ts";
+import { IntrospectedBaseClass } from "../gir/introspected-classes.ts";
 import { TwoKeyMap } from "../util.ts";
 import { 
     TypeIdentifier,
